+++ packages/data_layer_firestore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Adds `raw` method to Firestore source for untyped merges.
+
 ## 0.0.1
 
 * Adds initial FirestoreSource implementation.