+++ packages/data_layer_firestore_admin/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Adds `raw` method to Firestore source for untyped merges.
+
 ## 0.0.1
 
 - Initial release of `pkg:data_layer_firestore_admin`.